FSL Plaus commit 07022025
Updates to correct penalty scores displayed in the html output.

This impacted the following files:
- resources/overall_plaus.xlsx
- calculate_plausibility.R

On a secondary note, made a minor change to include add_fcs and a recoding for dnk and pnta values in the HHS indicator, which were erroneously allowed in the MSNA template tool this year. Proper FCS, HHS, rCSI and LCSI variables are needed for the report to calculate plausibility, so these functions are needed if the dataset doesnt already have proper formatting.